neutron: rename db when upgrading a containerized undercloud
When the undercloud was not containerized, the neutron database name was called neutron.
When we upgrade to a containerized undercloud, the database name is called ovs_neutron.
We introduced MigrateLegacyNeutronDb (false by default in the service but true when
the undercloud is containerized) that will rename the database during host_prep_tasks.

Also, we'll make sure mariadb is stopped before running Puppet steps,
but only when we deploy a containerized undercloud and also when mariadb
was actually running. The tasks are idempotent and tested.

Change-Id: I009cd38f4d10bf3942c8f18f90c6a0fa50858219
Closes-Bug: #1753247