Increase size of Heat HeatMaxJsonBodySize
Most of the HTTP servers default max body size to 1MB, Apache, Nginx.
That default value is ok as long as users mostly get data from server
but when something needs to be uploaded, 1MB limit is often not enough.
This is what we see with Heat and some of the large templates.
It's not uncommon that WSGI apps have 10MB+ body size limit,
so with the current Heat default value we are still in low range sizes.

Original Heat template HeatMaxJsonBodySize default value was 1M and now
it will be 4M. There is aslo corresponding change in instack-undercloud
where we bumped 'max_template_size' to the same value.

Change-Id: I974f6d4adbc08f2960a164b90ebb784ca466e76a
Closes-Bug: 1752467