Default environment/services/* to docker
This updates the relevant environment/services templates so that
they also default to docker. Without this change users of these
services could end up with mismatched (unsupported) deployments.

Additionally, this picks environment/services-docker as the
source of truth when resolving collisions for moving those over
environment/services. So environment/services now contain all of
the containerized services' env files used to be located in
environment/services-docker. The latter will be deleted later to
prevent future desync of contents.

Co-authored-by: Bogdan Dobrelya <bdobreli@redhat.com>
Change-Id: I923731f46ea26582160a11d2dfe85792ab74110b