Remove default value for pacmaker passwords
The default values for the PcsdPassword and PacemakerRemoteAuthkey
parameters have been removed, as they did not result in a functioning
pacemaker installation. These values are instead generated by
tripleo-common, and in the cases where they are not (direct API), we
want to fail explicitly if they are not provided.

Change-Id: I37f7a30e593263da2c18d1b2dad89e780a172a51
Closes-Bug: #1753822
(cherry picked from commit d57bd297a9470fedb15224e215733e01a198b286)