Fixes ODL container failing to start due to missing etc config
The ODL etc directory was being overridden with an empty mount directory
before kolla_start copied the puppet-generated config files.  The
puppet-generated config files only include modified configuration files
and not all of the default config files.  Therefore ODL was missing
config files when it started so the container was constantly crashing.

This patch removes the unwanted mount erasing the /opt/opendaylight/etc
directory and moves the upgrade file to be created in puppet-generated,
which will be copied at kolla start time for upgrade.  The
puppet-generated dir is read-only, so the REST call to disable upgrade
flag in ODL will only disable it for the running instance.  Therefore we
have to use ansible to write the file again to disable it incase ODL is
rebooted.

Closes-Bug: 1755916

Change-Id: Ie380cc41ca50a294a2647d673f339d02111bf6b3
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 97173caf8f68695140d207c2ab88226fd86659dc)