Add validation for SoftwareConfig outputs
SoftwareConfig/StructuredConfig outputs aren't supported with
config-download given that Heat doesn't know what the output values will
since Ansible is applying all configuration after the stack is complete.

This validation will report a warning whenever it finds use of outputs on
these resource types.

After config-download is the default and the Heat driven method is no
longer supported, we can switch this warning to an error.

Change-Id: I44d5ee3bab3d05ab0a59261d15ea915c75b35713