Add the service_config_settings from nova-compute base
The docker nova-compute.yaml file does not have the
service_config_settings from the base file. So the
collectd and fluentd parameters defined are not got
when the service run containerized.

The patch get the service_config_settings from the base file and
use it on the when is containerized.

Closes-Bug: #1757066
Change-Id: I0b3567edd7530d1ae0a69d3ee25bd1442f14fb58
(cherry picked from commit a3297afea9fd03c495c280cac881218032786355)