Work around packaging issue in iptables-services.
When iptables-services is upgraded it restarts the iptables services
which breaks connectivity because it looses the current iptables
rules.

So we add another network workaround.  This time around the
iptables-services package.

The first target is the update process as this is where the problem
should arise, but as the framework for network workaround is already
is place, we cover upgrade as well.

Partial-Bug: #1758291

Change-Id: Ia2f94058bac6cf28b4bd425385ffd629555c9609