Do not create NetworkVlanID is the value is not defined
In case of Upgrade, the {{role}}NetworkVlanID is not defined
in network_data.yaml in Pike.
So when the jinja template generate the network_data.yaml
the parameter_defaults for the NetworkVlanID result to
empty parameters.
If no VlanID has been provided by user environment files,
then we need to used the default defined in network/config/

Change-Id: If119f3900eb60a217b8209d59fb9bc900107addb