Fixes Neutron certificate and key permissions
The Neutron UID is not static and may be different between the host and
neutron container.  Since we generate certificates and keys on the host
for neutron and then mount them in a container, it is highly likely the
container Neutron UID will not match the one used on the host to
generate the files and reading these files will fail in the container.

This patch modifies the permissions after the files are mounted in the
container to be owned by the correct Neutron UID.

Closes-Bug: 1759049

Depends-On: I83b14b91d1ee600bd9d5863acba34303921368ce

Change-Id: Ibad3f1af4b44459e96a6dc9937e5fcef3e6335f4
Signed-off-by: Tim Rozet <trozet@redhat.com>
(cherry picked from commit 16731819c5bb92d9bb33c6fb8086a6a776bdef8c)