get-occ-config.sh: fix handling multiple nodes in a role
When using get-occ-config.sh with a role with a count greater
than 1, the script will now configure all nodes that are of that role
type instead of exiting after only configuring the first.

Previously, the expression ((i++)) evaluated to 0, and when using
arithmetic expansion, causes a rc of 1, which caused the script to exit
after configuring the first node.

Change-Id: I0a39f2b8f1de5302fd426999acfe5043b0361987
Closes-Bug: #1759581