Revert "FFU: Add support for panko service upgrade"
This reverts commit bd48087520c5f0846363bdc0c025508ba450ceb3.

After further inspection It seems that panko dbsync shouldn't be
needed, as it will upgrade an newly created empty db.

And this is assuming we find a way to:
 - configure panko database connection properly
 - create the db

Knowing that we don't have access to this information[1] as the
new hieradata haven't been rendered at this stage.

So all that to upgrade a newly (I guess empty) database seems like too
much trouble.

The db will be created in the last step of the FFU.

[1] https://github.com/openstack/tripleo-heat-templates/blob/stable/ocata/puppet/services/panko-base.yaml#L39..L75

Change-Id: Ie68849a7033c199c339d28cdb10c3dba9419904b
Closes-Bug: #1760135