masquerade: stop using YAQL for iptables data
See https://bugs.launchpad.net/tripleo/+bug/1760211 but the YAQL + Heat
format hasn't helped us to build the data needed by Puppet to create
IPtables rules for masquerading.

We'll solve it in puppet-tripleo with native hash iterations, so for
that we just export MasqueradeNetworks to Hiera and we'll use this data
from the puppet module directly.

Depends-On: I8dda3c164de90954855979529de4f1100a858b45

Related-Bug: #1760211
Change-Id: I81379cf93f505fb65c1ad7e6a2adcc6942b04bc0