undercloud: increase token expiration time
We did it in the past (3 years ago!) in instack-undercloud:
https://github.com/openstack/instack-undercloud/commit/43e792c6844d4a7081b718d7f89b0c40f5cfb708
in the context of: https://bugzilla.redhat.com/show_bug.cgi?id=1235908

This time, we have the same problem when the undercloud is
containeirized.
This patch is actually setting parity with keystone config from
instack-undercloud, but also raising an actual issue that will be
addressed this cycle.

In the meantime, let's increase the token expiration so we can move
forward with testing the containerized undercloud.

Change-Id: Iceaaf53fae44b5bcda9f6517f163939ba6be3d49
Related-Bug: #1761050