[queens] Put back (now empty) fluentd-client.yaml
This will allow us to get around a tricky error in heatclient template
processing with persisted user-environment.yaml on
upgrade. Interestingly, we need the old file to exist in the plan so
that we can map the resource to OS::Heat::None during the upgrade
process (heatclient tries to fetch it even though a subsequent env
file would map the resource to None), and all alternative solutions
are considerably worse than just putting the file back in place.

We will not need this file in Rocky as the resource will be mapped to
None before upgrading Q->R. This commit is specific for P->Q upgrade.

Change-Id: I39573cca4dc16a024582de562b850ecfa9bb1dfd
Co-Authored-By: Jose Luis Franco <jfrancoa@redhat.com>
Co-Authored-By: Sofer Athlan-Guyot <sathlang@redhat.com>
Closes-Bug: #1758406