Fixes the OvsVhostuserMode heat parameter
The mode was incorrectly defined as configuring the mode for OVS when it
is really configuring the client/server mode for QEMU.  A value of
'client' here will cause OVS to create the  vhostuser socket, and
actually make OVS the server.  This type of port is deprecated so the
default should be 'server' (QEMU creates the port).

Closes-Bug: 1762473

Change-Id: I7cd34b30f54766fe35e4a40a57b278166d0197cb
Signed-off-by: Tim Rozet <trozet@redhat.com>