Always run mysql init bundle
This init container runs docker-puppet manually and is responsible of
provisioning the mysql users and passwords. This currently doesn't get
ran every time since the configuration stays the same, even if the users
or passwords change (which are gotten from hieradata). Allowing this to
run every time will allow us to change database passwords

Closes-Bug: #1762991
Change-Id: I1f07272499b419079466cf9f395fb04a082099bd