Upgrade: make bundles use new container image name after upgrade
The major_upgrade tasks for HA services only allows to change the container
image tag used by bundles. It doesn't work when the image name changes.

Fix this unwanted behaviour by updating the bundle's attribute in pacemaker
to use container image <NEW>:pcmklatest instead of <CURRENT>:pcmklatest
We are constrained by the steps at when we can modify the bundle:
  . Image update must stay at step 3 when pacemaker is stopped.
  . image name used by the bundle must be available in docker when the
    bundle is updated

So we re-use the double tagging idiom to perform the image update:
  . At step 0, we tag the image pointed to by <CURRENT>:pcmklatest
    with an additional temporary tag <NEW>:pcmklatest.
    => this ensures that at step1, the new tag is available on all
       controller nodes.
  . At step 1, we update the resource bundle to use the new image
    name <NEW>:pcmklatest.
    => at the end of step1, the bundle will be configured with the
       new name, and be able to start even if the real container
       image hasn't be pulled yet.
  . At step 3, the existing code will download the real image
    <NEW>:<NEWTAG> and make tag <NEW>:pcmklatest point to it.

Since the bundle is always modified, we now stop and restart the
bundle resources unconditionally.

Also, move the mariadb upgrade task to step 3, when pacemaker is
guaranteed to be stopped, because the task assumes that no mysql
is running while it runs. Fix the mysql permission after rpm
upgrade on the host.

Change-Id: Ic87a66753b104b9f15db70fdccbd66d88cef94df
Closes-Bug: #1763001