NFV: Support for config-download to deploy node with kernel args
For NFV deployments, specific kernel args should be applied and
the nodes should be restarted before running the NetworkDeployment.
It is supported in the heat deployment via PreNetworkConfig. In the
config-download mechanism, ansible steps need to be improved
to handle the reboot and wait for the node.
Related-Bug: #1813610

Change-Id: I43b383ad0e04b8be6c321f8c5b05e628b2520141
(cherry picked from commit a3e4a9063612a617105e318e422d90706e4ed43e)