[NEWTON ONLY] Adjust NeutronAllowL3AgentFailover to new default.
The default value of default NeutronAllowL3AgentFailover has changed
to true.  But in Mitaka that was set to false in the registry.  So
that value persists and overrides the new default.

This has no impact currently as this value is ignored neutron as the
l3 agents are in HA configuration (this parameter only work when
non-HA).

But, it's still good to have for two reasons:

 - it simplifies recovery if we hit bug #1763322, see[1]
 - it will not cause unexpected issue if we offer the possibility to
   have non-HA l3 agent.

[1] https://bugzilla.redhat.com/attachment.cgi?id=1421308

Change-Id: Ibba351ff625abfe8fd5b1f76b9a49bac120cbd8a
Related-Bug: #1763322