Reduce verbosity during mysql bootstrap
During the initial deployment, a one-time container is used to bootstrap the
mysql databse on disk, create the required users and set their password. The
script that runs that is too verbose and logs some credentials in the
container's logs and in the journal.

Use kolla_extend_start directly instead of kolla_start to stop tracing shell
commands and reduce logging to the bare minimum for troubleshooting.

Closes-Bug: #1765339

Change-Id: I90827feff0d1b9fd8badb72e68e4c8dd8db8aea5