Add OVNCMSOptions in dvr environment files
This option must be enabled in a controller to be selected for hosting
a gateway router. But it is missing in DVR environment files. So when
these files are used for the deployment, vms were not able to reach
external network. So adding this option for DVR files.

Change-Id: Ibfc9baf3aca55b49cc197758cbf5a0df9daa8b9b
Closes-Bug: 1766183
(cherry picked from commit 39ef030553e65ed80326afcf339029ed0f2ae8e4)