Add volumes list of docker-puppet container when merging config volumes
When config volumes of same types are merged, the volumes to be mounted
for the continer is not merged. In order to run neutron container with
ovs agent, '/run/openvswitch' directory mount is required to execute
ovs-vsctl commands. This patch merges the volumes list of the same
config volumes instead of taking only the first one.

Closes-Bug: #1766759
Change-Id: I70972ad2fc339c9316befd74dda8555982f23316