Do not rely on defaults for DockerCephDaemonImage in CI.
When using mixed-versions deployment, as
it's the case for upgrades CI jobs. The
docker image version is being set by the
undercloud's tripleo-common package release.

As the undercloud is one release over the
overcloud, we end up deploying next's release
ceph docker image (ceph luminous in pike). By
setting the DockerCephDaemon parameter to the
right ceph docker image version we ensure that
the right version is being deployed

Closes-Bug: #1767329
Change-Id: Iff84601195722429f0e31334fee4845da0ca549c