Execute ceph-ansible with its own ansible.cfg
Modify the ceph-ansible execution so that it uses the ansible.cfg
that is shipped with ceph-ansible. ceph-ansible's own testing uses
this file and the project's maintainers recommend it be used.

This configuration file also uses ControlMaster=auto which is not
used by config download at this time. Adding this setting to the
ceph-ansible execution should make it take less time.

Change-Id: I774132eb8b7a43c921c0696635d56b4136301346
Closes-Bug: 1767252