Remove step_config from CinderVolume backend services
Cinder backends should not repeat the step_config that the CinderVolume
service uses to trigger things in puppet-tripleo. This is benign when
Cinder runs on baremetal, but causes problems when Cinder runs in
containers. It causes the cinder-volume service to run twice, in both
the container as well as on the baremetal host.

Closes-Bug: #1768063
Change-Id: Ie02553f74563883094a13c67a958b15371e90424
(cherry picked from commit 94b5c1155e7b8ccd54f0e17e8fb07f5a7a378aea)