Add basics for standalone node
For a standalone all-in-one, we need to create a basic role that has
some of the services, a network config for a single node and an
environment file that has all the services defined but disabled so
that we can enable just the services we will need. In the future, we
will likely make the service list more dynamic but for now it contains a
minimal set of services for a keystone/openshift/kubernetes deployment.

Change-Id: Ieb7c94563bd0132393b5fa268d743981f6e0b6f2
Related-Blueprint: all-in-one