Unnessary deploying of bootstrap-osd ceph auth entry
Ceph jewel creates the client.bootstrap-osd auth entry when the cluster
is created. heat does not need to pass this key to ceph::key as the
auth will already be populated on the monitors and the OSD nodes will
extract a copy from memory. This patch removes client.bootstrap-osd from
the ceph::profile::params::client_keys array.

Change-Id: Ia9c50afd3a88b27e6e0dca2ce150f91633681466
Closes-bug: 1769722
Depends-on: I11f134bed59da706a03ce4a70d941145c16f8175
Depends-On: https://review.opendev.org/#/c/712501/