Add ability to pre-assign IPs by role on ctlplane
This change adds the ability to pre-assign IP addresses
that will be used on the ctlplane network for each node.
The functionality is similar to the existing ips-from-pool
templates, but the IP will be passed to the Nova server
resource rather than a dedicated Neutron port (as happens
with the isolated networks templates). This allows for
compatibility with legacy installations for upgrades.
In testing, it also appears that the fixed IP can be
changed in a stack update, but more testing will have to
be done. Note that if an IP address is defined for
some nodes but not others, nodes without a fixed IP
will get an IP assigned automatically by Neutron.

Change-Id: I67513f54a60f5a50a2bc435099fbb2a643adc277
(cherry picked from commit 393476fda3652637bf6e5344f8815b3bb5398900)