Add acl to paths that are shared among related neutron processes
Neutron runs several processes to support features such as routers,
dhcp, etc. Neutron monitors these processes through shared domain
sockets, pids files, etc in /var/lib/neutron.  When upgrading a node
from a non-containerized deployment to containers, the user and group
ids are updated to that of the container and the supporting processes
are no longer able to access the shared files. This patch adds ACL to
the shared files to resolve this.

Note this only impacts upgraded nodes and only routers, networks, etc,
that existed before the upgrade.

Closes-Bug: #1772072
Change-Id: I8f28032488c30b5d38d8c7fa76872a2c3a642717