rerun *_init_bundles all the time
In the same spirit as change I1f07272499b419079466cf9f395fb04a082099bd
we want to rerun all pacemaker _init_bundles all the time. For a few main
reasons:
1) We will eventually support scaling-up roles that contain
   pacemaker-managed services and we need to rerun _init_bundles so that
   pacemaker properties are created for the newly added nodes.
2) When you replace a controller the pacemaker properties will be
   recreated for the newly added node.
3) We need to create appropriate iptables rules whenever we add a
   service to an existing deployment.

We do this by adding the DeployIdentifier to the environment so that
paunch will retrigger a run at every redeploy.

Partial-Bug: #1775196
Change-Id: Ifd48d74507609fc7f4abc269b61b2868bfbc9272