Don't use crudini to get rhn server
crudini is not installed by default on Red Hat derivitive OS's, so we
shouldn't rely on it. We probably can't just install it since this
script is the rhel-registration script, so there are likely no repos
enabled when we need it.

Instead just use grep/cut/sed to get the value we need.

Change-Id: I78fce8b6c7f1d3528f9d8c02772f95cb8ad3b3c8
Closes-Bug: #1771830
(cherry picked from commit b4ff453f87610988b7f76688bdba7e006ec91d20)