Add support of shared staging location for glance-direct
In the case of an HA setup (ie with 3 controller nodes) if image
import method is 'glance-direct', it would require a shared
consistent staging location on the controller nodes otherwise it
might happen that the image is uploaded to one node and triggering
the import is executed on a different node - where the image does not
exist.

Hence when import method 'glance-direct' is needed, operators need to
enable it explicitly with enabling NFS across the controller nodes
to share the staging.

Change-Id: I9a5bff5e5d0c3e109b73d691cc9088904a0401ac
Closes-Bug: #1765439