Set default application for Ceph Luminous openstack_pools
Luminous can assign application types to pools and ceph-ansible
supports this, provided it is passed an appropriate list of maps.
The list of maps is defined in openstack_pools and this change
updates this variable to contain the correct default application
type.

Includes update to environments/disable-telemetry.yaml so that
the GnocchiRbdPool is not created if telemetry is disabled.

Change-Id: I3966a8a1b170202e020684a36ecece78599296b5
Closes-Bug: #1772743