Add stack name to env() for OS::TripleO::WorkflowSteps
In I4b576a6e7fbfb18fa13221e2d080bf7876a8303e state information
will be persisted in Swift and the name of the Swift container
should be a function of the Heat stack in case multiple stacks
are deployed. This patch passes the name of the Heat stack to
the Mistral environment so that the workflow may access the
Heat stack name and name the Swift container accordingly.

Change-Id: I995ad32345a39238ffb9cbcf9966dedc60c75ff8
Related-Bug: #1769769