Use user specified private key when deleting temp keys
We use the user specified private key when we inject the temporary ssh
key, and we should also use it when we delete the temprorary ssh keys.
Even though this script is no longer used on master, we should commit
this so we can backport to queens.

Change-Id: Id389984b87209f959e4cf439abd8165a3a08ead3
Closes-Bug: #1773980
(cherry picked from commit d04e7ea4f3fce0c787c3eed6f1840b402435113c)