Generate and mount wrappers for haproxy in OVN metadata agent
OVN metadata agent uses haproxy as part of its implementation.
Running it in a separate container prevents dataplane breakages
(ie. restarting VMs or spawning new ones) on agent restart/stop.
This patch triggers the creation of such sidecar container and
mounting of haproxy wrapper for spawning it in a separate
container.

Change-Id: I59e08384080cda0b6c0f03c9ed8fb6f6a5661e6b
Related-Bug: #1749209
Signed-off-by: Daniel Alvarez <dalvarez@redhat.com>