Conditionally Enable Nuage VRS
Using this change, users can conditionally enable and
configure Nuage VRS on Overcloud nodes

In Pike, as neutron-server is still running as a service on
the hypervisor, this approach will configure VRS which is also
running on the hypervisor.

This review is meant for Pike and earlier releases only.

Change-Id: Ib563842f182383a9cec49ee57943039346098b53
Depends-On: I6a996a6b9e1b2d4c74e25e2582699eb9e023093a
Closes-Bug: 1776998
(cherry picked from commit 885f64daf86e6ff80340b05eec893cdf9ec98b4a)